@@ -5,32 +5,28 @@
   "description": "tingyuan's blog, https://tingyuan.me/nblog/",
   "scripts": {
     "dev": "vite",
-    "build:ssr": "cross-env VUE_SSR=true vite build --ssr",
-    "prebuild": " node public/buildblog.js",
-    "postbuild": "node public/prerender.js",
-    "build": "vite build && yarn build:ssr"
+    "build": "vue-tsc --noEmit && vite build",
+    "serve": "vite preview",
+    "prebuild": "vite build --ssr src/server.ts"
   },
   "dependencies": {
     "github-markdown-css": "^4.0.0",
-    "marked": "^1.2.7",
+    "marked": "^2.0.3",
     "normalize.css": "^8.0.1",
     "prismjs": "^1.23.0",
-    "vue": "^3.0.5"
+    "vue": "^3.0.11",
+    "vue-router": "4"
   },
   "devDependencies": {
-    "@types/marked": "^1.2.1",
+    "@types/marked": "^2.0.2",
     "@types/node": "^14.14.25",
-    "@types/prismjs": "^1.16.2",
-    "@vitejs/plugin-vue": "^1.1.4",
-    "@vue/compiler-sfc": "^3.0.5",
-    "@vue/server-renderer": "^3.0.5",
-    "@vuedx/typecheck": "^0.6.3",
-    "@vuedx/typescript-plugin-vue": "^0.6.3",
-    "cross-env": "^7.0.3",
+    "@types/prismjs": "^1.16.5",
+    "@vue/compiler-sfc": "^3.0.11",
+    "@vue/server-renderer": "^3.0.11",
     "domino": "^2.1.6",
-    "node-fetch": "^2.6.1",
-    "rimraf": "^3.0.2",
-    "typescript": "^4.1.3",
-    "vite": "^2.0.0-beta.65"
+    "typescript": "^4.2.4",
+    "@vitejs/plugin-vue": "^1.2.1",
+    "vue-tsc": "^0.0.24",
+    "vite": "^2.3.2"
   }
 }
