@@ -1,19 +1,42 @@
 <template>
   <transition-group name="blog-list" class="blog-list" tag="ul" v-if="list.length">
-    <li v-for="post of list" :key="post.id" class="blog-title">
-      <route-link :to="post.id">
-        {{post.title}}
-        <span style="font-size: .8em;">({{post.date}})</span>
-      </route-link>
+    <li v-for="name of list" :key="name" class="blog-title">
+      <router-link :to="'/' + props.cate + '/' + name" v-if="props.cate">
+        {{ name }}
+      </router-link>
+      <router-link v-else :to="'/' + name" class="blog-title">
+        {{name.split('/')[1]}}
+      </router-link>
     </li>
   </transition-group>
+  <div v-else-if='error'>{{error}}</div>
   <div v-else> Loading... </div>
 </template>
 
 <script lang="ts" setup>
-import { computed } from 'vue'
-import store from '../store'
-const list = computed(() => store.blogList)
+import { defineProps, watch, ref } from 'vue'
+import { getBlogArticleList, getAllBlogList } from '../blog'
+
+const props = defineProps({
+  cate: String
+})
+const list = ref<string[]>([])
+const error = ref('')
+watch(() => props.cate, (c) => {
+  error.value = ''
+  if (!c) {
+    list.value = getAllBlogList()
+  } else {
+    const articles = getBlogArticleList(c)
+    if (articles) {
+      list.value = articles
+    } else {
+      error.value = '请检查地址'
+    }
+  }
+}, {
+  immediate: true
+})
 
 </script>
 
